НебольшиÐе изменения в класс новой почты
